type problem 6

+++ src/datasets/semantic_dataset.py
@@ -48,7 +48,7 @@ class SemanticDataset(Dataset):
                 new_label[i] = np.where(label > 0, 1, 0)
         label = new_label
         image = np.expand_dims(image, axis=0)
-        image = th.from_numpy(image).to(th.float64)
+        image = th.from_numpy(image).to(th.float32)
         label = th.from_numpy(label).to(th.long)
